refactored sections list and sections status to constants and added static validation for VS Code
