+++ frontend/.vscode/settings.json
@@ -1,0 +1,12 @@
+{
+    "files.eol": "\n",
+    "javascript.referencesCodeLens.enabled": true,
+    "javascript.validate.enable": true,
+    "editor.suggest.localityBonus": true,
+    "eslint.autoFixOnSave": true,
+    "eslint.alwaysShowStatus": true,
+    "eslint.packageManager": "yarn",
+    "flow.enabled": false,
+    "flow.stopFlowOnExit": true,
+    "flow.showStatus": true
+}
