+++ frontend/src/components/editor/EditorSection.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import * as _ from 'lodash-es';
 import { Icon } from 'patternfly-react';
-import { EDITOR_STATUS } from '../../pages/operatorBundlePage/bundlePageUtils';
+import { EDITOR_STATUS } from '../../utils/constants';
 
 const EditorSection = ({ sectionStatus, title, description, sectionLocation, history }) => {
   const status = _.get(sectionStatus, sectionLocation);
