+++ frontend/src/redux/actions/editorActions.js
@@ -1,12 +1,21 @@
 import { reduxConstants } from '../index';
 import { clearAutosavedOperatorData } from '../../utils/operatorUtils';
 
+/**
+ * Update status of single section
+ * @param {EditorSectionNames} section
+ * @param {EDITOR_STATUS} status
+ */
 export const setSectionStatusAction = (section, status) => ({
   type: reduxConstants.SET_EDITOR_SECTION_STATUS,
   section,
   status
 });
 
+/**
+ * Update status of multiple sections at once
+ * @param {Record<EditorSectionNames, EDITOR_STATUS>} status
+ */
 export const setBatchSectionsStatusAction = status => ({
   type: reduxConstants.SET_EDITOR_ALL_SECTIONS_STATUS,
   status
@@ -25,11 +34,19 @@ export const resetEditorOperatorAction = () => {
   };
 };
 
+/**
+ * Updates operator package
+ * @param {OperatorPackage} operatorPackage
+ */
 export const updateOperatorPackageAction = operatorPackage => ({
   type: reduxConstants.SET_EDITOR_PACKAGE,
   operatorPackage
 });
 
+/**
+ * Update uploads list
+ * @param {UploadMetadata} uploads
+ */
 export const setUploadsAction = uploads => ({
   type: reduxConstants.SET_EDITOR_UPLOADS,
   uploads
