+++ frontend/src/utils/editorTypes.js
@@ -1,0 +1,11 @@
+/**
+ * @typedef {'metadata'|'owned-crds'|'required-crds'|'deployments'|'permissions'|'cluster-permissions'|'install-modes'|'package'} EditorSectionNames
+ */
+
+/**
+ * @typedef {object} EDITOR_STATUS
+ * @prop {'empty'} empty
+ * @prop {'pending'} pending
+ * @prop {'complete'} complete
+ * @prop {'errors'} errors
+ */
